requirements.txt
sounddevice==0.4.6
numpy==1.24.3
scipy==1.10.1
openai==1.3.0
pyperclip==1.8.2
python-dotenv==1.0.0 